Convert storage of calibration parameters from double-precision to single-precision in the ingest pipeline (the caltask) they are converted to single precision anyway, so no benifit to storing them in double-precision..
git-svn-id: https://svn.atnf.csiro.au/askapsdp/Src/trunk@3899 1db04d4c-0ff0-4329-ae0b-98c546476c6a

+++ Code/Components/CP/cpdataservices/server/current/src/askap/cp/calds/persist/BandpassSolutionElementBean.java
@@ -31,12 +31,12 @@ public class BandpassSolutionElementBean {
 	private short itsBeamID;
 	private int itsChan;
 	
-	private double itsG1Real;
-	private double itsG1Imag;
+	private float itsG1Real;
+	private float itsG1Imag;
 	private boolean itsG1Valid;
 	
-	private double itsG2Real;
-	private double itsG2Imag;
+	private float itsG2Real;
+	private float itsG2Imag;
 	private boolean itsG2Valid;
 	
 	/**
@@ -54,8 +54,8 @@ public class BandpassSolutionElementBean {
 	
 	public BandpassSolutionElementBean(long solutionID,
 			short antennaID, short beamID, int chan,
-			double g1Real, double g1Imag, boolean g1Valid,
-			double g2Real, double g2Imag, boolean g2Valid) {
+			float g1Real, float g1Imag, boolean g1Valid,
+			float g2Real, float g2Imag, boolean g2Valid) {
 		itsSolutionID = solutionID;
 		itsAntennaID = antennaID;
 		itsBeamID = beamID;
@@ -127,28 +127,28 @@ public class BandpassSolutionElementBean {
 	/**
 	 * @return the g1Real
 	 */
-	public double getG1Real() {
+	public float getG1Real() {
 		return itsG1Real;
 	}
 
 	/**
 	 * @param g1Real the g1Real to set
 	 */
-	public void setG1Real(double g1Real) {
+	public void setG1Real(float g1Real) {
 		itsG1Real = g1Real;
 	}
 
 	/**
 	 * @return the g1Imag
 	 */
-	public double getG1Imag() {
+	public float getG1Imag() {
 		return itsG1Imag;
 	}
 
 	/**
 	 * @param g1Imag the g1Imag to set
 	 */
-	public void setG1Imag(double g1Imag) {
+	public void setG1Imag(float g1Imag) {
 		itsG1Imag = g1Imag;
 	}
 
@@ -169,28 +169,28 @@ public class BandpassSolutionElementBean {
 	/**
 	 * @return the g2Real
 	 */
-	public double getG2Real() {
+	public float getG2Real() {
 		return itsG2Real;
 	}
 
 	/**
 	 * @param g2Real the g2Real to set
 	 */
-	public void setG2Real(double g2Real) {
+	public void setG2Real(float g2Real) {
 		itsG2Real = g2Real;
 	}
 
 	/**
 	 * @return the g2Imag
 	 */
-	public double getG2Imag() {
+	public float getG2Imag() {
 		return itsG2Imag;
 	}
 
 	/**
 	 * @param g2Imag the g2Imag to set
 	 */
-	public void setG2Imag(double g2Imag) {
+	public void setG2Imag(float g2Imag) {
 		itsG2Imag = g2Imag;
 	}
